Switch from Stream to ValueNotifier
